refactor(core): SMI-4813 split SkillParser.ts under 500-line gate
Extract the four exported types into `SkillParser.types.ts` (83 lines)
and the internal `parseYamlFrontmatter` helper into
`SkillParser.helpers.ts` (101 lines). `SkillParser.ts` slims from 500 →
343 lines, comfortably under the audit:standards 500-line gate.

`SkillParser.ts` re-exports the four types from `SkillParser.types.ts`,
so the public surface in `packages/core/src/indexer/index.ts` is
unchanged. `parseYamlFrontmatter` stays internal — only consumed by the
class itself; no test imported it directly.

Verification:
- Typecheck green (`npx tsc --build packages/core`)
- 96 tests pass across `GitHubIndexer.test.ts`,
  `EdgeCases.test.ts`, `check-references.test.ts`
- audit:standards file-length warning resolved (5 → 3 warnings;
  remaining all in-scope under SMI-4814/4815/4816)

Refs SMI-4812.
